Rename files Create Project folder structure Create Separate files for adapters and entities Rename Fragments
