+++ app/build.gradle
@@ -8,12 +8,12 @@ plugins {
 
 android {
     namespace 'jp.co.yumemi.android.code_check'
-    compileSdk 31
+    compileSdk 33
 
     defaultConfig {
         applicationId "jp.co.yumemi.android.codecheck"
         minSdk 23
-        targetSdk 31
+        targetSdk 33
         versionCode 1
         versionName "1.0"
