+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/activities/MainActivity.kt
@@ -1,12 +1,13 @@
 /*
  * Copyright © 2021 YUMEMI Inc. All rights reserved.
  */
-package jp.co.yumemi.android.code_check
+package jp.co.yumemi.android.code_check.ui.activities
 
 import androidx.appcompat.app.AppCompatActivity
+import jp.co.yumemi.android.code_check.R
 import java.util.*
 
-class TopActivity : AppCompatActivity(R.layout.activity_top) {
+class MainActivity : AppCompatActivity(R.layout.activity_top) {
 
     companion object {
         lateinit var lastSearchDate: Date
