+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/adapters/GitRepoListAdapter.kt
@@ -1,0 +1,40 @@
+package jp.co.yumemi.android.code_check.ui.adapters
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.RecyclerView
+import jp.co.yumemi.android.code_check.R
+import jp.co.yumemi.android.code_check.model.GitHubRepo
+import jp.co.yumemi.android.code_check.ui.fragments.diff_util
+
+class GitRepoListAdapter(
+    private val itemClickListener: OnItemClickListener,
+) : ListAdapter<GitHubRepo, GitRepoListAdapter.ViewHolder>(diff_util){
+
+    class ViewHolder(view: View): RecyclerView.ViewHolder(view)
+
+    interface OnItemClickListener{
+        fun itemClick(item: GitHubRepo)
+    }
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder
+    {
+        val _view= LayoutInflater.from(parent.context)
+            .inflate(R.layout.layout_item, parent, false)
+        return ViewHolder(_view)
+    }
+
+    override fun onBindViewHolder(holder: ViewHolder, position: Int)
+    {
+        val _item= getItem(position)
+        (holder.itemView.findViewById<View>(R.id.repositoryNameView) as TextView).text=
+            _item.name
+
+        holder.itemView.setOnClickListener{
+            itemClickListener.itemClick(_item)
+        }
+    }
+}
