+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/fragments/GitHubRepoDetailsFragment.kt
@@ -1,30 +1,29 @@
 /*
  * Copyright © 2021 YUMEMI Inc. All rights reserved.
  */
-package jp.co.yumemi.android.code_check
+package jp.co.yumemi.android.code_check.ui.fragments
 
 import android.os.Bundle
-import android.util.Log
 import android.view.View
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.navArgs
 import coil.load
-import jp.co.yumemi.android.code_check.TopActivity.Companion.lastSearchDate
-import jp.co.yumemi.android.code_check.databinding.FragmentTwoBinding
+import jp.co.yumemi.android.code_check.R
+import jp.co.yumemi.android.code_check.databinding.FragmentGitHubDetailsBinding
 
-class TwoFragment : Fragment(R.layout.fragment_two) {
+class GitHubRepoDetailsFragment : Fragment(R.layout.fragment_git_hub_details) {
 
-    private val args: TwoFragmentArgs by navArgs()
+    private val args: GitHubRepoDetailsFragmentArgs by navArgs()
 
-    private var binding: FragmentTwoBinding? = null
+    private var binding: FragmentGitHubDetailsBinding? = null
     private val _binding get() = binding!!
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        Log.d("検索した日時", lastSearchDate.toString())
+//        Log.d("検索した日時", lastSearchDate.toString())
 
-        binding = FragmentTwoBinding.bind(view)
+        binding = FragmentGitHubDetailsBinding.bind(view)
 
         var item = args.item
