+++ app/src/main/res/layout/fragment_git_hub_details.xml
