+++ app/src/main/res/navigation/nav_graph.xml
@@ -3,26 +3,26 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/apk/res-auto"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/oneFragment">
+    app:startDestination="@id/gitHubRepoListFragment">
 
     <fragment
-        android:id="@+id/oneFragment"
-        android:name="jp.co.yumemi.android.code_check.OneFragment"
+        android:id="@+id/gitHubRepoListFragment"
+        android:name="jp.co.yumemi.android.code_check.ui.fragments.GitHubRepoListFragment"
         android:label="@string/app_name"
-        tools:layout="@layout/fragment_one">
+        tools:layout="@layout/fragment_git_hub_repo_list">
         <action
             android:id="@+id/action_repositoriesFragment_to_repositoryFragment"
-            app:destination="@id/twoFragment" />
+            app:destination="@id/gitHubRepoDetailsFragment" />
     </fragment>
 
     <fragment
-        android:id="@+id/twoFragment"
-        android:name="jp.co.yumemi.android.code_check.TwoFragment"
+        android:id="@+id/gitHubRepoDetailsFragment"
+        android:name="jp.co.yumemi.android.code_check.ui.fragments.GitHubRepoDetailsFragment"
         android:label="@string/app_name"
-        tools:layout="@layout/fragment_two">
+        tools:layout="@layout/fragment_git_hub_details">
         <argument
             android:name="item"
-            app:argType="jp.co.yumemi.android.code_check.item" />
+            app:argType="jp.co.yumemi.android.code_check.model.GitHubRepo" />
     </fragment>
 
 </navigation>
